Fix UB caused by left shift of negative number

+++ src/lib/arch/aarch64/Instruction_execute.cc
@@ -2947,13 +2947,15 @@ void Instruction::execute() {
     }
     case Opcode::AArch64_MOVNWi: {  // movn wd, #imm{, LSL #shift}
       uint8_t shift = metadata.operands[1].shift.value;
-      uint32_t value = ~(metadata.operands[1].imm << shift);
+      uint32_t value =
+          ~(static_cast<uint64_t>(metadata.operands[1].imm) << shift);
       results[0] = static_cast<uint64_t>(value);
       break;
     }
     case Opcode::AArch64_MOVNXi: {  // movn xd, #imm{, LSL #shift}
       uint8_t shift = metadata.operands[1].shift.value;
-      uint64_t value = ~(metadata.operands[1].imm << shift);
+      uint64_t value =
+          ~(static_cast<uint64_t>(metadata.operands[1].imm) << shift);
       results[0] = value;
       break;
     }
